Pianta/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==40.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.3
django-allauth==0.53.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-resized==1.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.6
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.15
